net/iavf: fix devargs parsing
This patch fixes unknown parsing results in the absence of
devarg "no-poll-on-link-down" and fixes the devarg variable type.

Fixes: 5b3124a0a6ef ("net/iavf: support no polling when link down")
Fixes: 3e6a5d2d310a ("net/iavf: add devargs to enable VF auto-reset")
Cc: stable@dpdk.org

Acked-by: Qi Zhang <qi.z.zhang@intel.com>

+++ drivers/net/iavf/iavf.h
@@ -307,8 +307,8 @@ struct iavf_devargs {
 	uint8_t proto_xtr[IAVF_MAX_QUEUE_NUM];
 	uint16_t quanta_size;
 	uint32_t watchdog_period;
-	uint8_t  auto_reset;
-	uint16_t no_poll_on_link_down;
+	int auto_reset;
+	int no_poll_on_link_down;
 };
 
 struct iavf_security_ctx;
